+++ swanlab/cloud/__init__.py
@@ -8,7 +8,7 @@ r"""
     云端日志资源上传部分
 """
 from .start_thread import ThreadPool
-from .upload_types import UploadType
+from .task_types import UploadType
 from .dog.log_sniffer import LogSnifferTask
 
 __all__ = ["UploadType", "LogSnifferTask", "ThreadPool"]
